added text size control
