+++ Cargo.toml
@@ -21,3 +21,10 @@ rayon = "1.9.0"
 rodio = "0.17.3"
 serde = { version = "1.0.197", features = ["derive"] }
 spectrum-analyzer = "1.5.0"
+
+# [profile.release]
+# lto = true
+# opt-level = 3
+# codegen-units = 1
+# incremental = false
+
